Whitespace; RFC2822/RFC5321; testing thereof
Was missing trailing whitespace support in the LHS regexp, failed to
account for a `[FWS]` in the grammar.

Took a look at RFC5321 vs RFC2822's address grammar.  5321 simplifies
the rules, but also excludes some control characters.

So define both sets of patterns and use build-tag constrained helpers to
alias the active set to a const used in the patterns.

New tests for the trailing whitespace; new tests for the control
characters in RFC2822-only mode.

Travis supports build matrix in Go by Go version and by env; so set a
`BUILDTAGS` variable in the matrix and override the script to pass in
the tags.