Clarify the goveralls message re failures
I blinked at the test status, which seemed to be saying that there had
been a failure which had been ignored.  On review, the intent was to
note that any failures would be ignored.  Rephrased slightly in the hope
that the next time I go a year between looks at this output I might
confuse myself less.

[ci skip]